关联issue #18 and fixed #18
